--- src/RssItem.ts
@@ -1,7 +1,0 @@
-export interface RssItem {
-    title?: string;
-    pubDate?: string;
-    image?: string;
-    description?: string;
-    link?: string;
-}
