+++ src/index.html
@@ -48,7 +48,7 @@
                     <a href="#" download="feedData.json" id="exportButton" class="button-style">Export JSON</a>
 
                     <button id="button-add-feed" class="button-style">Add Feed</button>
-                  
+
                     <button id="add-feed-button" class="button-style">Add Feed</button>
                     <button id="upload-button" class="button-style">Upload</button>
                     <button id="download-button" class="button-style">Download</button>
@@ -82,7 +82,6 @@
                 <h2 class="expanded-article-title">Selected Article</h2>
                 <p class="expanded-article-content">This is the content of the selected article.</p>
             </div>
-
         </div>
     </body>
 </html>
